refactor(dep-005c): delete schemas.workflow bridge — migrate remaining callers to canonical imports
Complete the schemas.workflow migration trilogy (dep-005a/005b/005c).
Migrate 25+ files from deprecated schemas.workflow imports to canonical
paths (schemas.outputs, schemas.results, schemas.verification,
schemas.config). Delete all 19 bridge files and the workflow/ directory.
