--- src/karenina/schemas/workflow/judgment_results.py
@@ -1,9 +1,0 @@
-"""JudgmentResults class for collating and analyzing deep judgment results.
-
-DEPRECATED: Import from `karenina.schemas.results` instead.
-"""
-
-# Re-export from new location for backward compatibility
-from ..results.judgment import JudgmentResults
-
-__all__ = ["JudgmentResults"]
