--- src/karenina/schemas/workflow/rubric_dataframe.py
@@ -1,9 +1,0 @@
-"""DataFrame builders for rubric evaluation results.
-
-DEPRECATED: Import from `karenina.schemas.dataframes` instead.
-"""
-
-# Re-export from new location for backward compatibility
-from ..dataframes.rubric import RubricDataFrameBuilder
-
-__all__ = ["RubricDataFrameBuilder"]
