--- src/karenina/schemas/workflow/rubric_judgment_results.py
@@ -1,9 +1,0 @@
-"""RubricJudgmentResults class for detailed deep judgment rubric analysis.
-
-DEPRECATED: Import from `karenina.schemas.results` instead.
-"""
-
-# Re-export from new location for backward compatibility
-from ..results.rubric_judgment import RubricJudgmentResults
-
-__all__ = ["RubricJudgmentResults"]
