--- src/karenina/schemas/workflow/rubric_results.py
@@ -1,9 +1,0 @@
-"""RubricResults class for collating and analyzing rubric evaluation results.
-
-DEPRECATED: Import from `karenina.schemas.results` instead.
-"""
-
-# Re-export from new location for backward compatibility
-from ..results.rubric import RubricResults
-
-__all__ = ["RubricResults"]
