--- src/karenina/schemas/workflow/template_dataframe.py
@@ -1,9 +1,0 @@
-"""DataFrame builders for template verification results.
-
-DEPRECATED: Import from `karenina.schemas.dataframes` instead.
-"""
-
-# Re-export from new location for backward compatibility
-from ..dataframes.template import TemplateDataFrameBuilder
-
-__all__ = ["TemplateDataFrameBuilder"]
