--- src/karenina/schemas/workflow/template_results.py
@@ -1,9 +1,0 @@
-"""TemplateResults class for collating and analyzing template verification results.
-
-DEPRECATED: Import from `karenina.schemas.results` instead.
-"""
-
-# Re-export from new location for backward compatibility
-from ..results.template import TemplateResults
-
-__all__ = ["TemplateResults"]
