--- src/karenina/schemas/workflow/verification/job.py
@@ -1,9 +1,0 @@
-"""Verification job model.
-
-DEPRECATED: Import from `karenina.schemas.verification` instead.
-"""
-
-# Re-export from new location for backward compatibility
-from ...verification.job import VerificationJob
-
-__all__ = ["VerificationJob"]
