--- src/karenina/schemas/workflow/verification/result.py
@@ -1,9 +1,0 @@
-"""Verification result model.
-
-DEPRECATED: Import from `karenina.schemas.verification` instead.
-"""
-
-# Re-export from new location for backward compatibility
-from ...verification.result import VerificationResult
-
-__all__ = ["VerificationResult"]
