--- src/karenina/schemas/workflow/verification/result_components.py
@@ -1,21 +1,0 @@
-"""Component classes for verification results.
-
-DEPRECATED: Import from `karenina.schemas.verification` instead.
-"""
-
-# Re-export from new location for backward compatibility
-from ...verification.result_components import (
-    VerificationResultDeepJudgment,
-    VerificationResultDeepJudgmentRubric,
-    VerificationResultMetadata,
-    VerificationResultRubric,
-    VerificationResultTemplate,
-)
-
-__all__ = [
-    "VerificationResultMetadata",
-    "VerificationResultTemplate",
-    "VerificationResultRubric",
-    "VerificationResultDeepJudgment",
-    "VerificationResultDeepJudgmentRubric",
-]
