--- src/karenina/schemas/workflow/verification_result_set.py
@@ -1,9 +1,0 @@
-"""VerificationResultSet class - main container for verification results.
-
-DEPRECATED: Import from `karenina.schemas.results` instead.
-"""
-
-# Re-export from new location for backward compatibility
-from ..results.verification_result_set import VerificationResultSet
-
-__all__ = ["VerificationResultSet"]
