+++ tests/integration/test_deep_judgment_rubric_real_api.py
@@ -14,15 +14,14 @@ from datetime import UTC, datetime
 
 import pytest
 
-from karenina.schemas.verification.model_identity import ModelIdentity
-from karenina.schemas.workflow import (
+from karenina.schemas.results import RubricJudgmentResults, RubricResults
+from karenina.schemas.verification import (
     VerificationResult,
     VerificationResultDeepJudgmentRubric,
     VerificationResultMetadata,
     VerificationResultRubric,
 )
-from karenina.schemas.workflow.rubric_judgment_results import RubricJudgmentResults
-from karenina.schemas.workflow.rubric_results import RubricResults
+from karenina.schemas.verification.model_identity import ModelIdentity
 
 RubricJudgmentResults.model_rebuild()
